bge-m3 PCC: split large-batch sweep into test_model_multibatch.py
The nightly CI dispatch (test_ci_dispatch.py) runs the entire test_model.py
file. With batch 1/8/16/32 x 9 seq lengths that was 36 forward builds + the
batch32 trace replay -- ~31 min, risking CI timeout.

Move the batch 8/16/32 sweep and the batch32 trace-replay test into a new
test_model_multibatch.py. test_model.py now collects only the 9 batch-1
cases (the CI-facing set); the large-batch sweep is run explicitly via
test_model_multibatch.py. No coverage lost, identical bodies/tolerances
(PCC_THRESHOLD=0.94).
