+++ models/demos/wormhole/bge_m3/tests/pcc/test_model.py
@@ -15,8 +15,6 @@ from models.demos.wormhole.bge_m3.tests.test_utils import (
 from models.demos.wormhole.bge_m3.tt.common import create_tt_model
 
 MODEL_ID = "BAAI/bge-m3"
-BATCH_SIZE_B32 = 32
-SEQ_LEN_B32 = 512
 PCC_THRESHOLD = 0.94
 
 
@@ -34,14 +32,10 @@ def model_artifacts(model_location_generator):
     return backbone, state_dict, model_id_or_path
 
 
-@pytest.mark.slow
-@pytest.mark.parametrize("seq_len", SEQUENCE_LENGTHS, ids=[f"S{s}" for s in SEQUENCE_LENGTHS])
-@pytest.mark.parametrize("batch_size", [1, 8, 16, 32], ids=["batch1", "batch8", "batch16", "batch32"])
-def test_model_full_end_to_end(device, model_artifacts, batch_size, seq_len, reset_seeds):
-    """End-to-end HF-vs-TT PCC for batch 1/8/16/32 x all sequence lengths (bf8_b).
+def _run_full_end_to_end(device, model_artifacts, batch_size, seq_len):
+    """Shared body: end-to-end HF-vs-TT PCC for one (batch_size, seq_len), bf8_b.
 
-    Gated at PCC_THRESHOLD=0.94. Filter combos with -k, e.g.
-    `-k "S512"`, `-k "batch1"`, or `-k "batch8 and S512"`.
+    Gated at PCC_THRESHOLD=0.94.
     """
     require_single_device(device)
     backbone, state_dict, model_id_or_path = model_artifacts
@@ -84,68 +78,13 @@ def test_model_full_end_to_end(device, model_artifacts, batch_size, seq_len, res
 
 
 @pytest.mark.slow
-def test_model_full_end_to_end_batch32_seq512_trace_replay(device, model_artifacts, reset_seeds):
-    require_single_device(device)
-    backbone, state_dict, model_id_or_path = model_artifacts
+@pytest.mark.parametrize("seq_len", SEQUENCE_LENGTHS, ids=[f"S{s}" for s in SEQUENCE_LENGTHS])
+def test_model_full_end_to_end(device, model_artifacts, seq_len, reset_seeds):
+    """End-to-end HF-vs-TT PCC for batch 1 x all sequence lengths (bf8_b).
 
-    model_args, tt_model, _ = create_tt_model(
-        mesh_device=device,
-        max_batch_size=BATCH_SIZE_B32,
-        max_seq_len=SEQ_LEN_B32,
-        dtype=ttnn.bfloat8_b,
-        state_dict=state_dict,
-        hf_model_name=model_id_or_path,
-    )
-    torch.manual_seed(42)
-    input_ids = torch.randint(
-        low=0,
-        high=model_args.vocab_size,
-        size=(BATCH_SIZE_B32, SEQ_LEN_B32),
-        dtype=torch.long,
-    )
-    non_pad_token_id = (int(model_args.pad_token_id) + 1) % model_args.vocab_size
-    input_ids[input_ids == model_args.pad_token_id] = non_pad_token_id
-    token_type_ids = torch.zeros_like(input_ids)
-    attention_mask = None
-
-    with torch.no_grad():
-        reference_output = (
-            backbone(
-                input_ids=input_ids,
-                attention_mask=attention_mask,
-                token_type_ids=token_type_ids,
-                position_ids=None,
-                return_dict=True,
-            )
-            .last_hidden_state.unsqueeze(1)
-            .to(torch.float32)
-        )
-
-    tt_input_ids = to_ttnn_ids(input_ids, device)
-    tt_token_type_ids = to_ttnn_ids(token_type_ids, device)
-    trace_captured = False
-    try:
-        # Compile once outside trace capture; the trace should contain only the steady-state forward.
-        tt_model.forward(
-            input_ids=tt_input_ids,
-            attention_mask=attention_mask,
-            token_type_ids=tt_token_type_ids,
-        )
-        ttnn.synchronize_device(device)
-
-        tt_model.capture_trace(
-            input_ids=tt_input_ids,
-            attention_mask=attention_mask,
-            token_type_ids=tt_token_type_ids,
-            position_ids=None,
-            mesh_device=device,
-            cq_id=0,
-        )
-        trace_captured = True
-        tt_output = tt_model.execute_trace(blocking=False, synchronize=True)
-        tt_output_torch = to_torch(tt_output, expected_shape=(BATCH_SIZE_B32, 1, SEQ_LEN_B32, model_args.dim))
-    finally:
-        if trace_captured:
-            tt_model.release_trace()
-
-    assert_pcc(reference_output, tt_output_torch, PCC_THRESHOLD)
+    This is the CI-facing variant: batch 1 only (9 runs) to stay within the CI
+    time budget. The larger-batch sweep lives in
+    `test_model_full_end_to_end_multibatch`. Gated at PCC_THRESHOLD=0.94.
+    Filter combos with -k, e.g. `-k "S512"`.
+    """
+    _run_full_end_to_end(device, model_artifacts, batch_size=1, seq_len=seq_len)
