+++ models/demos/wormhole/bge_m3/tests/pcc/test_model_multibatch.py
@@ -1,0 +1,167 @@
+# SPDX-FileCopyrightText: © 2026 Tenstorrent USA, Inc.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Large-batch end-to-end PCC sweep for BgeM3Model (batch 8/16/32).
+
+Split out of ``test_model.py`` so the nightly CI dispatch (which runs the whole
+``test_model.py`` file) only executes the batch-1 sweep and stays within its
+time budget. Run this file explicitly for the full multi-batch sweep:
+
+    pytest models/demos/wormhole/bge_m3/tests/pcc/test_model_multibatch.py
+
+Gated at PCC_THRESHOLD=0.94 (same as test_model.py). Filter combos with -k,
+e.g. `-k "S512"`, `-k "batch8"`, or `-k "batch8 and S512"`.
+"""
+
+import pytest
+import torch
+
+import ttnn
+from models.demos.wormhole.bge_m3.tests.test_utils import (
+    SEQUENCE_LENGTHS,
+    assert_pcc,
+    require_single_device,
+    to_torch,
+    to_ttnn_ids,
+)
+from models.demos.wormhole.bge_m3.tt.common import create_tt_model
+
+MODEL_ID = "BAAI/bge-m3"
+BATCH_SIZE_B32 = 32
+SEQ_LEN_B32 = 512
+PCC_THRESHOLD = 0.94
+
+
+@pytest.fixture(scope="module")
+def model_artifacts(model_location_generator):
+    transformers = pytest.importorskip("transformers")
+    model_id_or_path = str(model_location_generator(MODEL_ID, download_if_ci_v2=True, ci_v2_timeout_in_s=1800))
+    hf_model = transformers.AutoModelForCausalLM.from_pretrained(
+        model_id_or_path,
+        torch_dtype=torch.bfloat16,
+    ).eval()
+
+    backbone = hf_model.roberta if hasattr(hf_model, "roberta") else hf_model
+    state_dict = hf_model.state_dict()
+    return backbone, state_dict, model_id_or_path
+
+
+def _run_full_end_to_end(device, model_artifacts, batch_size, seq_len):
+    """Shared body: end-to-end HF-vs-TT PCC for one (batch_size, seq_len), bf8_b.
+
+    Gated at PCC_THRESHOLD=0.94.
+    """
+    require_single_device(device)
+    backbone, state_dict, model_id_or_path = model_artifacts
+
+    model_args, tt_model, _ = create_tt_model(
+        mesh_device=device,
+        max_batch_size=batch_size,
+        max_seq_len=seq_len,
+        dtype=ttnn.bfloat8_b,
+        state_dict=state_dict,
+        hf_model_name=model_id_or_path,
+    )
+    torch.manual_seed(42)
+    input_ids = torch.randint(low=0, high=model_args.vocab_size, size=(batch_size, seq_len), dtype=torch.long)
+    non_pad_token_id = (int(model_args.pad_token_id) + 1) % model_args.vocab_size
+    input_ids[input_ids == model_args.pad_token_id] = non_pad_token_id
+    token_type_ids = torch.zeros_like(input_ids)
+    attention_mask = None
+    with torch.no_grad():
+        reference_output = (
+            backbone(
+                input_ids=input_ids,
+                attention_mask=attention_mask,
+                token_type_ids=token_type_ids,
+                position_ids=None,
+                return_dict=True,
+            )
+            .last_hidden_state.unsqueeze(1)
+            .to(torch.float32)
+        )
+
+    tt_output = tt_model.forward(
+        input_ids=to_ttnn_ids(input_ids, device),
+        attention_mask=attention_mask,
+        token_type_ids=to_ttnn_ids(token_type_ids, device),
+    )
+    tt_output_torch = to_torch(tt_output, expected_shape=(batch_size, 1, seq_len, model_args.dim))
+
+    assert_pcc(reference_output, tt_output_torch, PCC_THRESHOLD)
+
+
+@pytest.mark.slow
+@pytest.mark.parametrize("seq_len", SEQUENCE_LENGTHS, ids=[f"S{s}" for s in SEQUENCE_LENGTHS])
+@pytest.mark.parametrize("batch_size", [8, 16, 32], ids=["batch8", "batch16", "batch32"])
+def test_model_full_end_to_end_multibatch(device, model_artifacts, batch_size, seq_len, reset_seeds):
+    """End-to-end HF-vs-TT PCC for batch 8/16/32 x all sequence lengths (bf8_b)."""
+    _run_full_end_to_end(device, model_artifacts, batch_size=batch_size, seq_len=seq_len)
+
+
+@pytest.mark.slow
+def test_model_full_end_to_end_batch32_seq512_trace_replay(device, model_artifacts, reset_seeds):
+    require_single_device(device)
+    backbone, state_dict, model_id_or_path = model_artifacts
+
+    model_args, tt_model, _ = create_tt_model(
+        mesh_device=device,
+        max_batch_size=BATCH_SIZE_B32,
+        max_seq_len=SEQ_LEN_B32,
+        dtype=ttnn.bfloat8_b,
+        state_dict=state_dict,
+        hf_model_name=model_id_or_path,
+    )
+    torch.manual_seed(42)
+    input_ids = torch.randint(
+        low=0,
+        high=model_args.vocab_size,
+        size=(BATCH_SIZE_B32, SEQ_LEN_B32),
+        dtype=torch.long,
+    )
+    non_pad_token_id = (int(model_args.pad_token_id) + 1) % model_args.vocab_size
+    input_ids[input_ids == model_args.pad_token_id] = non_pad_token_id
+    token_type_ids = torch.zeros_like(input_ids)
+    attention_mask = None
+
+    with torch.no_grad():
+        reference_output = (
+            backbone(
+                input_ids=input_ids,
+                attention_mask=attention_mask,
+                token_type_ids=token_type_ids,
+                position_ids=None,
+                return_dict=True,
+            )
+            .last_hidden_state.unsqueeze(1)
+            .to(torch.float32)
+        )
+
+    tt_input_ids = to_ttnn_ids(input_ids, device)
+    tt_token_type_ids = to_ttnn_ids(token_type_ids, device)
+    trace_captured = False
+    try:
+        # Compile once outside trace capture; the trace should contain only the steady-state forward.
+        tt_model.forward(
+            input_ids=tt_input_ids,
+            attention_mask=attention_mask,
+            token_type_ids=tt_token_type_ids,
+        )
+        ttnn.synchronize_device(device)
+
+        tt_model.capture_trace(
+            input_ids=tt_input_ids,
+            attention_mask=attention_mask,
+            token_type_ids=tt_token_type_ids,
+            position_ids=None,
+            mesh_device=device,
+            cq_id=0,
+        )
+        trace_captured = True
+        tt_output = tt_model.execute_trace(blocking=False, synchronize=True)
+        tt_output_torch = to_torch(tt_output, expected_shape=(BATCH_SIZE_B32, 1, SEQ_LEN_B32, model_args.dim))
+    finally:
+        if trace_captured:
+            tt_model.release_trace()
+
+    assert_pcc(reference_output, tt_output_torch, PCC_THRESHOLD)
